requirements.txt
blinker==1.6.2
cffi==1.15.1
click==8.1.7
colorama==0.4.6
cryptography==41.0.4
Flask==2.3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pycparser==2.21
PyJWT==2.8.0
Werkzeug==2.3.7
